+++ utility.h
@@ -28,6 +28,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 
 /* -------------------------------------------------------------------------- */
 #include <stdio.h>
+#include <stdint.h>
 
 
 /* Program name and version */